Interpolation and Integration

CompositeGrids.Interp — Module

Provide interpolation and integration.

CompositeGrids.Interp.integrate1D — Method

function integrate1D(::CompositeIntegrate, data, xgrid)

calculate integration of data[i] on xgrid call integrate1D for each subgrid and return the sum.

#Arguments:

xgrid: one-dimensional grid of x

data: one-dimensional array of data

CompositeGrids.Interp.integrate1D — Method

function integrate1D(::NoIntegrate, data, xgrid)

calculate integration of data[i] on xgrid works for grids that do not have integration weight stored

#Arguments:

xgrid: one-dimensional grid of x

data: one-dimensional array of data

CompositeGrids.Interp.integrate1D — Method

function integrate1D(::WeightIntegrate, data, xgrid)

calculate integration of data[i] on xgrid works for grids that have integration weight stored

#Arguments:

xgrid: one-dimensional grid of x

data: one-dimensional array of data

CompositeGrids.Interp.integrate1D — Method

function integrate1D(data, xgrid)

calculate integration of data[i] on xgrid

#Arguments:

xgrid: one-dimensional grid of x

data: one-dimensional array of data

CompositeGrids.Interp.interp1D — Method

function interp1D(::ChebInterp, data, xgrid, x)

linear interpolation of data(x), barycheb for BaryCheb grid

#Arguments:

xgrid: one-dimensional grid of x

data: one-dimensional array of data

x: x

CompositeGrids.Interp.interp1D — Method

function interp1D(::CompositeInterp,data, xgrid, x)

linear interpolation of data(x), first floor on panel to find subgrid, then call interp1D on subgrid

#Arguments:

xgrid: one-dimensional grid of x

data: one-dimensional array of data

x: x

CompositeGrids.Interp.interp1D — Method

function interp1D(::FloorInterp,data, xgrid, x)

linear interpolation of data(x), use floor and linear1D

#Arguments:

xgrid: one-dimensional grid of x

data: one-dimensional array of data

x: x

CompositeGrids.Interp.interp1D — Method

function interp1D(data, xgrid, x)

linear interpolation of data(x)

#Arguments:

xgrid: one-dimensional grid of x

data: one-dimensional array of data

x: x

CompositeGrids.Interp.interpGrid — Method

function interpGrid(::CompositeInterp, data, xgrid, grid)

linear interpolation of data(grid[1:end]), return a Vector grid should be sorted.

#Arguments:

xgrid: one-dimensional grid of x

data: one-dimensional array of data

grid: points to be interpolated on

CompositeGrids.Interp.interpGrid — Method

function interpGrid(::Union{FloorInterp,ChebInterp}, data, xgrid, grid)

linear interpolation of data(grid[1:end]), return a Vector simply call interp1D on each points

#Arguments:

xgrid: one-dimensional grid of x

data: one-dimensional array of data

grid: points to be interpolated on

CompositeGrids.Interp.interpGrid — Method

function interpGrid(data, xgrid, grid)

linear interpolation of data(grid[1:end]), return a Vector

#Arguments:

xgrid: one-dimensional grid of x

data: one-dimensional array of data

grid: points to be interpolated on

CompositeGrids.Interp.linear1D — Method

function linear1D(data, xgrid, x)

linear interpolation of data(x)

#Arguments:

xgrid: one-dimensional grid of x

data: one-dimensional array of data

x: x

CompositeGrids.Interp.linearND — Method

function linearND(data, xgrids, xs)

linear interpolation of data(xs)

#Arguments:

xgrids: n-dimensional grids, xgrids[i] is a 1D grid

data: n-dimensional array of data

xs: list of x, x[i] corresponds to xgrids[i]